- concentrated Torah readings by book; - cleanup.
