+++ dates/src/main/scala/org/podval/calendar/schedule/tanach/Tachanun.scala
@@ -2,6 +2,7 @@ package org.podval.calendar.schedule.tanach
 
 /*
 Days without Tachanun
+
 Рош ходеш
 Йом тов
 Хол- амоэд
@@ -22,6 +23,7 @@ Days without Tachanun
 Ту биав
 Все дни кроме Песах шени, кануна Рош ашана и кануна Йом кипура их канун в минху не говорят таханун.
 В Песах шени и шабат не читают таханун (в биркат амазон и шма аль амита) в ночь на их исходе
+Shabbos Mevorchim
 
 Хасидские даты
 10 кислева
